Export class instead of function
