--- packages/api-utility/src/position.ts
@@ -1,17 +1,0 @@
-/*
- * Only necessary in Electron and OpenFin as browser does not know about
- * more than 1 screen.
- */
-export const getPositionFromDisplay = (displayId: string, position: ssf.Position): Promise<ssf.Position> => {
-  if (!displayId) {
-    return Promise.resolve(position);
-  }
-
-  return ssf.Screen.getDisplays().then((displays) => {
-    const display = displays.filter(d => d.id === displayId)[0];
-    return {
-      x: display.bounds.x + position.x,
-      y: display.bounds.y + position.y
-    };
-  });
-};
